+++ src/mock.rs
@@ -1,21 +1,53 @@
 #[doc(hidden)]
-pub struct Mock(Box<dyn FnOnce(()) -> () + Send>);
+pub enum StoredMock {
+    Once(Box<dyn FnOnce(()) -> () + Send>),
+    Many(Box<dyn FnMut(()) -> () + Send>, MockTimes),
+}
 
-impl std::fmt::Debug for Mock {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        f.write_str("mock")
+#[derive(Debug)]
+pub enum MockTimes {
+    Always,
+    Times(usize),
+}
+
+impl MockTimes {
+    pub fn decrement(&mut self) {
+        if let MockTimes::Times(times) = self {
+            *times -= 1;
+        }
     }
 }
 
-impl Mock {
-    pub(crate) unsafe fn new<I, O>(mock: impl FnOnce(I) -> O + Send) -> Self {
+#[doc(hidden)]
+pub enum ReturnedMock<'a> {
+    Once(Box<dyn FnOnce(()) -> () + Send>),
+    Many(&'a mut (dyn FnMut(()) -> () + Send)),
+}
+
+impl std::fmt::Debug for StoredMock {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            StoredMock::Once(_) => f.write_str("once mock"),
+            StoredMock::Many(_, count) => write!(f, "mock {:?} times", count),
+        }
+    }
+}
+
+impl StoredMock {
+    pub(crate) unsafe fn once<I, O>(mock: impl FnOnce(I) -> O + Send) -> Self {
         let mock = Box::new(mock) as Box<dyn FnOnce(_) -> _>;
         let mock = std::mem::transmute(mock);
-        Mock(mock)
+        StoredMock::Once(mock)
+    }
+
+    pub(crate) unsafe fn many<I, O>(mock: impl FnMut(I) -> O + Send, times: MockTimes) -> Self {
+        let mock = Box::new(mock) as Box<dyn FnMut(_) -> _>;
+        let mock = std::mem::transmute(mock);
+        StoredMock::Many(mock, times)
     }
 }
 
-impl Mock {
+impl ReturnedMock<'_> {
     /// # Safety
     ///
     /// [#[methods]] makes sure this function is called correctly when
@@ -23,7 +55,16 @@ impl Mock {
     ///
     /// [#\[methods\]]: methods
     pub unsafe fn call<I, O>(self, input: I) -> O {
-        let mock: Box<dyn FnOnce(I) -> O> = std::mem::transmute(self.0);
-        mock(input)
+        match self {
+            ReturnedMock::Once(mock) => {
+                let mock: Box<dyn FnOnce(I) -> O> = std::mem::transmute(mock);
+                mock(input)
+            }
+            ReturnedMock::Many(mock) => {
+                let mock = &mut *(mock as *mut (dyn std::ops::FnMut(()) + std::marker::Send)
+                    as *mut dyn std::ops::FnMut(I) -> O);
+                mock(input)
+            }
+        }
     }
 }