+++ src/mock_store.rs
@@ -1,5 +1,5 @@
-use crate::Mock;
-use std::collections::HashMap;
+use crate::mock::{MockTimes, ReturnedMock, StoredMock};
+use std::collections::{self, HashMap};
 
 #[doc(hidden)]
 /// ```
@@ -31,7 +31,7 @@ impl<T> MaybeFaux<T> {
 #[derive(Debug, Default)]
 #[doc(hidden)]
 pub struct MockStore {
-    mocks: HashMap<&'static str, Mock>,
+    mocks: HashMap<&'static str, StoredMock>,
 }
 
 #[doc(hidden)]
@@ -53,7 +53,7 @@ impl MockStore {
         id: &'static str,
         mock: impl FnOnce(I) -> O + Send,
     ) {
-        self.mocks.insert(id, Mock::new(mock));
+        self.mocks.insert(id, StoredMock::once(mock));
     }
 
     pub fn mock_once<I, O>(&mut self, id: &'static str, mock: impl FnOnce(I) -> O + 'static + Send)
@@ -62,11 +62,75 @@ impl MockStore {
         O: 'static,
     {
         unsafe {
-            self.mocks.insert(id, Mock::new(mock));
+            self.mocks.insert(id, StoredMock::once(mock));
         }
     }
 
-    pub fn get_mock(&mut self, id: &str) -> Option<Mock> {
-        self.mocks.remove(id)
+    /// # Safety
+    ///
+    /// Do not use this function without going through [When].  It is
+    /// the caller's responsability to pass a mock that is safe.
+    ///
+    /// [When]: When
+    pub unsafe fn unsafe_mock<I, O>(
+        &mut self,
+        id: &'static str,
+        mock: impl FnMut(I) -> O + Send,
+        times: MockTimes,
+    ) {
+        self.mocks.insert(id, StoredMock::many(mock, times));
+    }
+
+    pub fn mock<I, O>(
+        &mut self,
+        id: &'static str,
+        mock: impl FnMut(I) -> O + 'static + Send,
+        times: MockTimes,
+    ) where
+        I: 'static,
+        O: 'static,
+    {
+        unsafe {
+            self.unsafe_mock(id, mock, times);
+        }
+    }
+
+    pub fn get_mock(&mut self, id: &'static str) -> Option<ReturnedMock> {
+        match self.mocks.entry(id) {
+            // no mock stored
+            collections::hash_map::Entry::Vacant(_) => None,
+            collections::hash_map::Entry::Occupied(mut entry) => match entry.get_mut() {
+                // we did not remove the mock on its "last" possible call so remove it now
+                // we do this because ReturnedMock::Many wants a reference
+                // so the mock must still live somewhere even during its last call
+                StoredMock::Many(_, MockTimes::Times(0)) => {
+                    entry.remove();
+                    None
+                }
+                // only a single mock
+                // remove and return mock
+                StoredMock::Once(_) => {
+                    let mock = entry.remove();
+                    match mock {
+                        StoredMock::Once(mock) => Some(ReturnedMock::Once(mock)),
+                        StoredMock::Many(_, _) => unreachable!(),
+                    }
+                }
+                // mock that can be called multiple times
+                // return mock but do not remove it
+                // call into_mut and stop using the get_mut because of lifetime shenanigans
+                StoredMock::Many(_, _) => match entry.into_mut() {
+                    StoredMock::Many(many, times) => {
+                        times.decrement();
+                        // One may think that doing &mut *many is the same thing
+                        // but NO IT IS NOT Why? ¯\_(ツ)_/¯
+                        // Be very careful because I do not understand why
+                        // and the other way fails some tests... but not all?
+                        Some(ReturnedMock::Many(many.as_mut()))
+                    }
+                    _ => unreachable!(),
+                },
+            },
+        }
     }
 }